feat: FCM 알림 서비스 및 이벤트 처리 구현
- FCM 토큰 저장 API 구현
- 모델 학습 상태(완료/실패) 알림 전송 기능 구현
- 카프카 이벤트 리스너 및 발행자 구현
- 모델 학습 완료/실패 처리 로직 개선

+++ spring/roll_model/src/main/java/com/ccc/roll_model/fcm/application/FCMService.java
@@ -50,8 +50,8 @@ public class FCMService {
 
     // 모델 학습 상태에 따른 메시지 발송
     @Transactional
-    public void sendModelTrainingStatusNotification(Long memberId, String status, String modelName) {
-        List<FCMTokenEntity> tokens = fcmTokenRepository.findByMemberIdAndIsActiveTrue(memberId);
+    public void sendModelTrainingStatusNotification(Integer memberId, String status, String modelName) {
+        List<FCMTokenEntity> tokens = fcmTokenRepository.findByMemberIdAndIsActiveTrue(memberId.intValue());
         if (tokens.isEmpty()) {
             log.info("No active FCM tokens found for member: {}", memberId);
             return; // 활성화된 토큰이 없으면 발송 불가
